Moved queue to cmd/collector/app.queue
